Invalid javascript header, not complaint with the embedded IANA reference
As described in http://www.ietf.org/rfc/rfc4329.txt application/x-javascript was temporary but not final and therefore this should be changed, while (new) embedded (tv) browsers do support IANA standards, but could perhaps not-know/skip application/x-javascript support which will mean that Javascript will not be executed.

I do understand that many plugins depend on this reference (like the Jenkins S3 plugin), but this is invalid. Any input about use-cases that could be impacted?